pyproject.toml
[project]
name = "otk"
version = "0.1.0"
description = "ecDNA analysis tool based on deep learning"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "tqdm>=4.65",
    "click>=8.1",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]

[project.scripts]
otk = "otk.cli:cli"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
otk = ["configs/*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
